refactor(server): set every single default values inside validation
